refactor(admin-ui): split UserDetailPage.tsx into focused sub-components
UserDetailPage.tsx (861 lines, 10 useState, 11 useQuery, 14 useMutation
all in one component) becomes a thin orchestrator: it owns only the
primary user fetch, the delete-user flow, and the impersonation flow, and
renders five section components that each own their own queries/mutations:

- UserProfileSection: profile form + update mutation
- UserSecuritySection: MFA status/reset + password reset
- UserRolesSection: realm role mappings + client role mappings
- UserGroupsSection: group membership
- UserSessionsSection: active sessions + offline sessions

Add useFormSeed(entity, seedFn) to hooks/, replacing the useState +
comparison-on-render reseeding pattern that was duplicated between
UserDetailPage and components/realms/GeneralSettingsForm.tsx. Only
UserProfileSection (the file being touched here) is migrated to it;
GeneralSettingsForm's copy is left as-is since retrofitting every existing
occurrence is out of scope for this issue.

No useUserDetailQueries shared read-layer was added: the codebase's
established split pattern (components/realms/*Form.tsx, used by the
already-split RealmDetailPage) has each section own its queries directly,
and a shared read layer would partially undo that.

UserDetailPage had zero existing test coverage; added
UserDetailPage.test.tsx (8 tests) covering rendering, form seeding, empty
states, the delete-confirmation flow, and a 404 case, since a refactor
this size needs a regression net.

Closes #1298
